+++ src/p_saveg.h
@@ -69,7 +69,8 @@ typedef enum saveg_compat_e
   saveg_nugget200,
   saveg_nugget210,
   saveg_nugget300, // Identified by "Nugget 2.4.0"
-  saveg_current, // saveg_nugget320
+  saveg_nugget320,
+  saveg_current, // saveg_nugget330
 } saveg_compat_t;
 
 extern saveg_compat_t saveg_compat;