+++ src/p_setup.c
@@ -54,6 +54,9 @@
 #include "w_wad.h"
 #include "z_zone.h"
 
+// [Nugget]
+#include "m_array.h"
+
 //
 // MAP related Lookup tables.
 // Store VERTEXES, LINEDEFS, SIDEDEFS, etc.
@@ -427,6 +430,8 @@ void P_LoadThings (int lump)
   int  i, numthings = W_LumpLength (lump) / sizeof(mapthing_t);
   byte *data = W_CacheLumpNum (lump,PU_STATIC);
 
+  int *enemies = NULL, numenemies = 0; // [Nugget]
+
   for (i=0; i<numthings; i++)
     {
       mapthing_t *mt = (mapthing_t *) data + i;
@@ -456,9 +461,49 @@ void P_LoadThings (int lump)
       mt->options = SHORT(mt->options);
 
       P_SpawnMapThing (mt);
+
+      // [Nugget] ------------------------------------------------------------
+
+      if (x2monsters)
+      {
+        const mobjtype_t mobjtype = P_FindDoomedNum(mt->type);
+
+        if (mobjtype < num_mobj_types
+            && ((mobjinfo[mobjtype].flags & MF_COUNTKILL) || mobjtype == MT_SKULL))
+        {
+          array_push(enemies, i);
+          numenemies++;
+        }
+      }
     }
 
-  // [Nugget] Reset milestones / ---------------------------------------------
+  // [Nugget] /===============================================================
+
+  // Custom Skill: duplicate monster spawns ----------------------------------
+
+  if (x2monsters)
+  {
+    P_ToggleDuplicateSpawns(true);
+
+    for (i = 0;  i < numenemies;  i++)
+    {
+      mapthing_t *mt = (mapthing_t *) data + enemies[i];
+
+      mt->x = SHORT(mt->x);
+      mt->y = SHORT(mt->y);
+      mt->angle = SHORT(mt->angle);
+      mt->type = SHORT(mt->type);
+      mt->options = SHORT(mt->options);
+
+      P_SpawnMapThing (mt);
+    }
+
+    array_clear(enemies);
+
+    P_ToggleDuplicateSpawns(false);
+  }
+
+  // Reset milestones --------------------------------------------------------
 
   if (totalkills) { complete_milestones &= ~MILESTONE_KILLS; }
   else            { complete_milestones |=  MILESTONE_KILLS; }
@@ -466,7 +511,7 @@ void P_LoadThings (int lump)
   if (totalitems) { complete_milestones &= ~MILESTONE_ITEMS; }
   else            { complete_milestones |=  MILESTONE_ITEMS; }
 
-  // [Nugget] ---------------------------------------------------------------/
+  // [Nugget] ===============================================================/
 
   Z_Free (data);
 }